+++ Client/Client.Initialize.cs
@@ -241,8 +241,9 @@ public partial class Client
         if (options.OptionEnabled(OptionsConstants.Comp.Vile, compLevelDef.CompLevel))
             compat.VileGhosts.Set(true, writeToConfig: false);
 
-        var id24skies = m_archiveCollection.Definitions.Id24SkyDefinition.Data.Skies;
-        if (id24skies != null && id24skies.Count > 0 && m_config.Render.SkyMode != SkyRenderMode.Vanilla)
+        var id24skies = m_archiveCollection.Definitions.Id24SkyDefinition.Data.Skies ?? [];
+        var id24mapping = m_archiveCollection.Definitions.Id24SkyDefinition.Data.FlatMapping ?? [];
+        if ((id24skies.Count > 0 || id24mapping.Count > 0) && m_config.Render.SkyMode != SkyRenderMode.Vanilla)
         {
             m_config.Render.SkyMode.Set(SkyRenderMode.Vanilla, writeToConfig: false);
             Log.Info("SKYDEFS: Sky render mode set to vanilla");