+++ RELEASENOTES.md
@@ -15,3 +15,4 @@
 - Fix incorrect warnings for sounds and invalid bex string memonic with custom sounds prefixed with USER_.
 - Fix id24 skies to have defaults set for when not defined outside of flatmapping and correct lookup for animations.
 - Fix UMAPINFO/MAPINFO EnterText/ExitText/SecretExitText/ExitText with escaped double quotes.
+- Fix setting vanilla sky render mode with skydefs when only flatmapping is defined with no skies.